+++ tests/functions.test.ts
@@ -1,12 +1,12 @@
 import HomeAssistantJavaScriptTemplates from '../src';
-import { HASS } from './constants';
+import { HOME_ASSISTANT_ELEMENT } from './constants';
 
 describe('Function tests', () => {
 
     let compiler: HomeAssistantJavaScriptTemplates;
     
     beforeEach(() => {
-        compiler = new HomeAssistantJavaScriptTemplates(HASS);
+        compiler = new HomeAssistantJavaScriptTemplates(HOME_ASSISTANT_ELEMENT);
     });
 
     const methods = [