+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [2.0.0] - 2024-01-28
+
+[BREAKING CHANGE]:
+    - The main class doesn‘t accept a `hass` object anymore. It needs an HTML element that contains the `hass` object as a property. This is to solve a bug that the templates were returining always the old version of the states.
+
 ## [1.2.0] - 2024-01-28
 
 - Rollback the `None` returns and return `undefined` in these cases.